style(UserProfile): conform better with style rfc
