--- src/UserProfile/constants.js
@@ -1,14 +1,0 @@
-export const NO_ACCESS_MSG = 'You have no access to storage service. Please contact an admin to get it!';
-export const NO_ACCESS_KEY = 'You don\'t have any access key. Please create one!';
-export const CONFIRM_DELETE_MSG = 'Are you sure you want to make this key inactive?';
-export const SECRET_KEY_MSG = 'This secret key is only displayed this time. Please save it!';
-export const DELETE_BTN = 'Delete';
-export const CREATE_USER_BTN = 'Create user access';
-export const CREATE_ACCESS_KEY_BTN = 'Create access key';
-export const ACCESS_KEY_COLUMN = 'Access key(s)';
-export const PROJECT_COLUMN = 'Project(s)';
-export const RIGHT_COLUMN = 'Right(s)';
-export const CREATE_DATE_COLUMN = 'Created date';
-export const STATUS_COLUMN = 'Status';
-export const LIST_PROJECT_MSG = 'You have access to the following project(s)';
-export const LIST_ACCESS_KEY_MSG = 'You have the following access key(s)';
