+++ src/index.js
@@ -16,12 +16,11 @@ import QueryNode, { clearResultAndQuery } from './QueryNode/ReduxQueryNode';
 import DataDictionary from './DataDictionary/ReduxDataDictionary';
 import DataDictionaryNode from './DataDictionary/ReduxDataDictionaryNode';
 import ProjectSubmission from './Submission/component';
-import UserProfile from './UserProfile/component';
+import UserProfile, { loginUserProfile, fetchAccess } from './UserProfile/ReduxUserProfile';
 import CertificateQuiz from './Certificate/ReduxQuiz';
 import GraphQLQuery, { fetchSchema } from './GraphQLEditor/ReduxGqlEditor';
 import { loginSubmissionAPI } from './Submission/actions';
 import { fetchDictionary } from './queryactions';
-import { loginUserProfile, fetchAccess } from './UserProfile/actions';
 import { app } from './localconf';
 import browserHistory from './history';
 import { theme } from './theme';
